+++ src/ConfigurationTrait.php
@@ -14,7 +14,8 @@ use ArrayAccess;
 /**
  * ConfigurationTrait which retrieves options from configuration
  *
- * Use this trait if you want to retrieve options from a configuration and optional to perform a mandatory option check
+ * Use this trait if you want to retrieve options from a configuration and optional to perform a mandatory option check.
+ * Default options are merged and overridden of the provided options.
  */
 trait ConfigurationTrait
 {
@@ -29,64 +30,73 @@ trait ConfigurationTrait
     abstract public function packageName();
 
     /**
-     * Returns options and it is possible to disable exceptions and return a default value instead.
-     *
-     * @see \Interop\Config\ObtainOptions::options
-     *
-     * @param $config
-     * @param bool $throwException Whether or not to throw an exception, optional default true
-     * @param mixed $default Value is returned of $throwException is false, optional default null
-     * @return mixed options
+     * @see \Interop\Config\HasConfig::canRetrieveOptions
      */
-    public function options($config, $throwException = true, $default = null)
+    public function canRetrieveOptions($config)
     {
         if (!is_array($config) && !$config instanceof ArrayAccess) {
-            if (false === $throwException) {
-                return $default;
-            }
+            return false;
+        }
+        $id = null;
+
+        if ($this instanceof HasContainerId) {
+            $id = $this->containerId();
+        }
+        $vendorName = $this->vendorName();
+        $packageName = $this->packageName();
+
+        if (!isset($config[$vendorName][$packageName])
+            || (null !== $id && !isset($config[$vendorName][$packageName][$id]))
+        ) {
+            return false;
+        }
+        $options = $config[$vendorName][$packageName];
+
+        if (null !== $id) {
+            $options = $options[$id];
+        }
+        return is_array($options) || $options instanceof ArrayAccess;
+    }
+
+    /**
+     * @see \Interop\Config\ObtainsOptions::options
+     */
+    public function options($config)
+    {
+        if (!is_array($config) && !$config instanceof ArrayAccess) {
             throw new Exception\InvalidArgumentException(
                 sprintf(
-                    '$config parameter provided to "%s" must be an "%s" or "%s"',
+                    '$config parameter provided to "%s" must either be of type "array" or implement "\ArrayAccess".',
                     __METHOD__,
                     'array',
                     'ArrayAccess'
                 )
             );
         }
-
         $id = null;
 
         if ($this instanceof HasContainerId) {
             $id = $this->containerId();
         }
-
         $vendorName = $this->vendorName();
         $packageName = $this->packageName();
 
         // this is the fastest way to determine a configuration error (performance)
         if (!isset($config[$vendorName][$packageName][$id])) {
-            if (false === $throwException) {
-                return $default;
-            }
             if (!isset($config[$vendorName][$packageName])) {
                 if (empty($config[$vendorName])) {
-                    throw new Exception\RuntimeException(
+                    throw new Exception\OutOfBoundsException(
                         sprintf('No vendor configuration "%s" available', $vendorName)
                     );
                 }
-                throw new Exception\OptionNotFoundException(sprintf(
-                    'No options set in configuration "' . "['%s']['%s']",
-                    $vendorName,
-                    $packageName
-                ));
+                throw new Exception\OptionNotFoundException(
+                    sprintf('No options set for configuration "' . "['%s']['%s']", $vendorName, $packageName)
+                );
             }
             if (null !== $id) {
-                throw new Exception\OptionNotFoundException(sprintf(
-                    'No options set in configuration "' . "['%s']['%s']['%s']",
-                    $vendorName,
-                    $packageName,
-                    $id
-                ));
+                throw new Exception\OptionNotFoundException(
+                    sprintf('No options set for configuration "' . "['%s']['%s']['%s']", $vendorName, $packageName, $id)
+                );
             }
         }
         $options = $config[$vendorName][$packageName];
@@ -94,11 +104,20 @@ trait ConfigurationTrait
         if (null !== $id) {
             $options = $options[$id];
         }
-        // check for mandatory options
+        if (!is_array($options) && !$options instanceof ArrayAccess) {
+            throw new Exception\UnexpectedValueException(
+                sprintf(
+                    'Options of configuration ' . "['%s']['%s']%s"
+                    . ' must either be of type "array" or implement "\ArrayAccess".',
+                    $this->vendorName(),
+                    $this->packageName(),
+                    $id ? '["' . $id . '""]' : ''
+                )
+            );
+        }
         if ($this instanceof HasMandatoryOptions) {
             $this->checkMandatoryOptions($this->mandatoryOptions(), $options);
         }
-        // check for default options
         if ($this instanceof HasDefaultOptions) {
             $options = array_replace_recursive($this->defaultOptions(), $options);
         }
@@ -134,7 +153,7 @@ trait ConfigurationTrait
                 $mandatoryOption,
                 $this->vendorName(),
                 $this->packageName(),
-                $id ? '[' . $id . ']' : ''
+                $id ? '["' . $id . '""]' : ''
             ));
         }
     }